LIBWEB-5393. Refactored Drupal database functionality into gateway class
Refactored functionality that communicates with the Drupal database into
a "DrupalGateway" class.

https://issues.umd.edu/browse/LIBWEB-5393
